streamlit_requirements.txt
streamlit
opencv-python-headless
pillow
numpy 
pytesseract
ultralytics
